refactor: util 함수 typescript 로 migration
