--- src/utils/index.js
@@ -1,27 +1,0 @@
-export const pareCardFormState = ({
-  firstCardNumber,
-  secondCardNumber,
-  thirdCardNumber,
-  fourthCardNumber,
-  ownerName,
-  secureCode,
-  expiredMonth,
-  expiredYear,
-  firstPassword,
-  secondPassword,
-  cardType,
-}) => {
-  return {
-    firstCardNumber,
-    secondCardNumber,
-    thirdCardNumber,
-    fourthCardNumber,
-    ownerName,
-    secureCode,
-    expiredMonth,
-    expiredYear,
-    firstPassword,
-    secondPassword,
-    cardType,
-  };
-};
